Kori-bot/requirements2.txt
absl-py==1.0.0
aiocache==0.11.1
aiocqhttp==1.4.2
aiofiles==0.8.0
aiohttp==3.8.1
aiorequests==0.1.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
APScheduler==3.8.1
arrow==1.2.2
asgiref==3.5.0
astunparse==1.6.3
async-asgi-testclient==1.4.9
async-timeout==4.0.2
asyncio-dgram==1.2.0
attrs==21.4.0
Automat==0.8.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beniget==0.4.1
binaryornot==0.4.4
bleach==4.1.0
blinker==1.4
build==0.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
# cloud-init==20.1
colorama==0.4.4
#command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cookiecutter==1.7.3
cryptography==36.0.1
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.28
databases==0.5.5
dblatex===0.3.11py3
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
docker==5.0.3
docopt==0.6.2
docutils==0.18.1
dominate==2.6.0
emoji==1.6.3
enchant==0.1.0
entrypoints==0.3
fastapi==0.73.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.3.0
fuzzywuzzy==0.18.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.12.0
h2==4.1.0
h5py==3.6.0
hpack==4.0.0
httpcore==0.14.7
httplib2==0.14.0
httptools==0.2.0
httpx==0.22.0
httpx-socks==0.7.3
hypercorn==0.13.2
hyperframe==6.0.1
hyperlink==19.0.0
idna==3.3
imageio==2.16.0
importlib-metadata==4.11.1
incremental==16.10.1
iniconfig==1.1.1
install==1.3.5
isc==2.0
itsdangerous==2.0.1
jaraco.context==4.1.1
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.0.3
jinja2-time==0.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
loguru==0.6.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mcstatus==8.0.0
more-itertools==8.12.0
multidict==5.2.0
nb==0.1.2
nb-cli==0.6.6
netifaces==0.10.4
networkx==2.7
nonebot-adapter-cqhttp==2.0.0b1
nonebot-adapter-onebot==2.0.0b1
nonebot-plugin-abbrreply==1.1.2
nonebot-plugin-analysis-bilibili==2.3.2
nonebot-plugin-apscheduler==0.1.2
nonebot-plugin-biliav==1.0.6
nonebot-plugin-caiyunai==0.2.2
nonebot-plugin-chatrecorder==0.1.5
nonebot-plugin-chess==0.1.0
nonebot-plugin-code==0.0.4
nonebot-plugin-datastore==0.3.0
nonebot-plugin-emojimix==0.1.3
nonebot-plugin-epicfree==0.1.7
nonebot-plugin-fortune==0.4.1
nonebot-plugin-gocqhttp==0.4.5
nonebot-plugin-heweather==0.5.2
nonebot-plugin-htmlrender==0.0.4.4
nonebot-plugin-ipypreter==0.1.5
nonebot-plugin-localstore==0.1.0
nonebot-plugin-logo==0.2.0
nonebot-plugin-memes==0.2.2.2
nonebot-plugin-morning==0.2.0a2
nonebot-plugin-petpet==0.2.4
nonebot-plugin-read-60s==0.2.3
nonebot-plugin-remake==0.2.4
nonebot-plugin-repeater==2.0.0
nonebot-plugin-roll==0.2.1
nonebot-plugin-russian==0.2.2
nonebot-plugin-setu==0.0.6
nonebot-plugin-songpicker2==0.1.7
nonebot-plugin-statistical==0.4.5
nonebot-plugin-status==0.3.0
nonebot-plugin-what2eat==0.2.3
nonebot-plugin-withdraw==0.2.0
nonebot-plugin-wordcloud==0.0.6
nonebot2==2.0.0b2
nonebug==0.2.1
numpy==1.22.2
oauthlib==3.2.0
opencv-doc-zh==2020.9.21.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pathlib==1.0.1
pep517==0.12.0
Pillow==9.0.1
pipreqs==0.4.11
pkginfo==1.8.2
playwright==1.19.0
pluggy==1.0.0
ply==3.11
poyo==0.5.0
priority==2.0.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.9.0
pyee==8.1.0
pyenchant==3.2.2
pyfiglet==0.8.post1
Pygments==2.11.2
PyGObject==3.36.0
pygtrie==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymdown-extensions==9.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pytest==6.2.5
pytest-asyncio==0.16.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-markdown-math==0.8
python-slugify==5.0.2
python-socks==2.0.3
pythran==0.11.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==6.0
Quart==0.14.1
random-cat-gif==1.0.0
readme-renderer==32.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
ssh-import-id==5.10
starlette==0.17.1
systemd-python==234
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.2.9
toml==0.10.2
tomli==2.0.1
tomlkit==0.9.2
tqdm==4.62.3
twine==3.8.0
Twisted==18.9.0
typing==3.7.4.3
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==20.3
ufw==0.36
ujson==5.1.0
unattended-upgrades==0.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
visitor==0.1.3
wadllib==1.3.3
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.1
Werkzeug==2.0.3
wolframalpha==5.0.0
wordcloud==1.8.1
wrapt==1.13.3
wsproto==1.0.0
xmltodict==0.12.0
yarg==0.1.9
yarl==1.7.2
zipp==3.7.0
zope.interface==4.7.1
